+++ src/calva-fmt/src/config.ts
@@ -1,15 +1,14 @@
 import * as vscode from 'vscode';
+import * as filesCache from '../../files-cache';
+import { cljfmtOptions } from '../../../out/cljs-lib/cljs-lib';
 
 function readConfiguration() {
     let workspaceConfig = vscode.workspace.getConfiguration("calva.fmt");
+    const cljfmtContent = filesCache.content(workspaceConfig.get("configPath"));
     return {
         "format-as-you-type": workspaceConfig.get("formatAsYouType") as boolean,
-        "indentation?": workspaceConfig.get("indentation"),
-        "remove-surrounding-whitespace?": workspaceConfig.get("removeSurroundingWhitespace"),
-        "remove-trailing-whitespace?": workspaceConfig.get("removeTrailingWhitespace"),
-        "insert-missing-whitespace?": workspaceConfig.get("insertMissingWhitespace"),
-        "remove-consecutive-blank-lines?": workspaceConfig.get("removeConsecutiveBlankLines"),
-        "align-associative?": workspaceConfig.get("alignMapItems")
+        "cljfmt-string": cljfmtContent,
+        "cljfmt-options": cljfmtOptions(cljfmtContent),
     };
 }