+++ test-data/test2.clj
@@ -1,15 +1,31 @@
+(a(b
+   (c
+     #f
+      (#b
+        [:f :b :z])
+     #z
+      1)))
+(comment
+  1
+  #foo
+   #bar
+    #baz
+     #{:foo (#foo '[1 2 3])})
+
 (comment
   
   #foo :bar
   (foo #foo ('foo^' :bar))
-  #_#_
-  'abc #foo  @~#(foo :bar) \space
+  #_ 'abc 
+  #foo  @~#(foo :bar) \space
   \' {:foo (#foo '[1 2 3])} 'abc
   #_#_
   #foo
-   'bar
+   #bar
+    #baz
+     #{:foo (#foo '[1 2 3])}
   #foo
-   #{:foo (#foo '[1 2 3])}
+   'bar
   \a [] "a"
 
   )